+++ lib/mix/tasks/gen.devices.ex
@@ -1,18 +1,31 @@
-defmodule Mix.Tasks.Gen.Devices do
-  @moduledoc false
+defmodule Mix.Tasks.NervesHub.Gen.Devices do
+  @moduledoc """
+  Generate a collection of devices for an Organization and Product.
+
+  ## Examples
+
+      mix nerves_hub.gen.devices my-org my-product 1000
+  """
+
+  @shortdoc "Generate a collection of devices"
 
   use Mix.Task
 
-  @shortdoc "Generate a mass of devices"
+  alias NervesHub.Accounts
+  alias NervesHub.Devices
+  alias NervesHub.Products
 
+  @requirements ["app.start"]
+  @preferred_cli_env :dev
+
+  @impl Mix.Task
   def run([org_name, product_name, count]) do
-    Mix.Task.run("app.start")
     count = String.to_integer(count)
-    {:ok, org} = NervesHub.Accounts.get_org_by_name(org_name)
-    {:ok, product} = NervesHub.Products.get_product_by_org_id_and_name(org.id, product_name)
+    {:ok, org} = Accounts.get_org_by_name(org_name)
+    {:ok, product} = Products.get_product_by_org_id_and_name(org.id, product_name)
 
     current_count =
-      NervesHub.Devices.get_device_count_by_org_id_and_product_id(org.id, product.id)
+      Devices.get_device_count_by_org_id_and_product_id(org.id, product.id)
 
     current_count..(current_count + count)
     |> Enum.map(fn i ->
@@ -23,7 +36,7 @@ defmodule Mix.Tasks.Gen.Devices do
       lng = -180..180 |> Enum.random()
       lat = -90..90 |> Enum.random()
 
-      NervesHub.Devices.create_device(%{
+      Devices.create_device(%{
         org_id: org.id,
         product_id: product.id,
         identifier: "generated-#{i}",
